+++ packages/ak-editor-button/src/index.js
@@ -3,6 +3,7 @@ import cx from 'classnames';
 import 'style!./host.less';
 import shadowStyles from './shadow.less';
 
+/* don't blur the prose mirror editor */
 function preventDefault(e) {
   e.preventDefault();
 }
@@ -12,12 +13,12 @@ export default define('ak-editor-button', {
     return (
       <div className={cx(shadowStyles.locals.root, { [shadowStyles.locals.active]: elem.active })}>
         <style>{shadowStyles.toString()}</style>
-        <button disabled={elem.disabled} onmousedown={preventDefault}><slot /></button>
+        <button disabled={elem.disabled} onMousedown={preventDefault}><slot /></button>
       </div>
     );
   },
   props: {
-    active: prop.boolean({ attribute: true, default: false }),
-    disabled: prop.boolean({ attribute: true, default: false }),
+    active: prop.boolean({ attribute: true }),
+    disabled: prop.boolean({ attribute: true }),
   },
 });