+++ packages/ak-editor-toolbar-block-type/src/index.js
@@ -1,12 +1,12 @@
-import { vdom, define, prop } from 'skatejs';
+import { vdom, define, prop, emit } from 'skatejs';
 import cx from 'classnames';
 import styles from './index.less';
 import FontSelect from './font-select';
 import Option from './option';
-import Overlay from './overlay';
+import Blanket from 'ak-blanket';
 
 const fonts = {
-  paragraph: 'Paragraph',
+  normalText: 'Normal text',
   heading1: 'Heading 1',
   heading2: 'Heading 2',
   heading3: 'Heading 3',
@@ -21,19 +21,29 @@ function selectFont(elem) {
   return (event) => {
     elem.selectedFont = event.detail.font;
     toggle(elem);
+
+    // remove this when reactify v0.0.7 is released (https://github.com/webcomponents/react-integration/commit/53f8bf59a76b0ea0929bf2e95866ce949456eef5)
+    emit(elem, 'selectfont', { detail: { font: elem.selectedFont } });
   };
 }
 
 export default define('ak-editor-toolbar-block-type', {
   render(elem) {
     return (
-      <div className={styles.locals.root}>
+      <div
+        className={styles.locals.root}
+        on-ak-blanket-click={() => {
+          toggle(elem);
+        }}
+        onSelectFont={selectFont(elem)}
+      >
         <style>{styles.toString()}</style>
-        <Overlay open={elem.dropdownOpen} ontoggleDropdown={() => toggle(elem)} />
+        {elem.dropdownOpen ? <Blanket clickable /> : null}
         <FontSelect
+          disabled={elem.disabled}
           className={styles.locals.fontSelect}
           selectedReadableName={fonts[elem.selectedFont]}
-          ontoggleDropdown={() => toggle(elem)}
+          onToggleDropdown={() => toggle(elem)}
           active={elem.dropdownOpen}
         >
           <ul
@@ -44,7 +54,6 @@ export default define('ak-editor-toolbar-block-type', {
             {Object.keys(fonts).map(font => (
               <li><Option
                 font={font}
-                onselectFont={selectFont(elem)}
                 active={elem.selectedFont === font}
               >{fonts[font]}</Option></li>
             ))}
@@ -56,5 +65,6 @@ export default define('ak-editor-toolbar-block-type', {
   props: {
     dropdownOpen: prop.boolean({ attribute: true }),
     selectedFont: prop.string({ attribute: true, default: Object.keys(fonts)[0] }),
+    disabled: prop.boolean({ attribute: true }),
   },
 });