+++ packages/ak-inline-dialog/src/index.js
@@ -1,10 +1,30 @@
 /** @jsx vdom */
-/* eslint react/no-unknown-property: 0 */
 import 'style!./host.less';
-import { enumeration } from 'akutil-common';
-import { vdom, prop, define } from 'skatejs';
+import { enumeration, KeyPressHandler } from 'akutil-common';
+import { vdom, prop, define, emit } from 'skatejs';
 import shadowStyles from './shadow.less';
-import Layer, { POSITION_ATTRIBUTE_ENUM, CONSTRAIN_ATTRIBUTE_ENUM } from 'ak-layer'; // eslint-disable-line no-unused-vars, max-len
+import Layer, { POSITION_ATTRIBUTE_ENUM, CONSTRAIN_ATTRIBUTE_ENUM } from 'ak-layer';
+import Blanket from 'ak-blanket';
+
+let keyPress;
+function closeDialog(elem) {
+  return () => {
+    elem.open = false;
+  };
+}
+
+function renderBlanketIfNeeded(elem) {
+  if (elem.hasBlanket) {
+    return (
+      <Blanket
+        tinted={elem.isBlanketTinted}
+        clickable={elem.isBlanketClickable}
+      />
+    );
+  }
+
+  return '';
+}
 
 /**
  * @description The definition for the InlineDialog component.
@@ -12,9 +32,16 @@ import Layer, { POSITION_ATTRIBUTE_ENUM, CONSTRAIN_ATTRIBUTE_ENUM } from 'ak-lay
  * @example @html <ak-inline-dialog target="#target"></ak-inline-dialog>
  * @example @js import InlineDialog from 'ak-inline-dialog';
  * const myDialog = new InlineDialog();
- *
  */
-const definition = {
+export default define('ak-inline-dialog', {
+  attached(elem) {
+    keyPress = new KeyPressHandler('ESCAPE', closeDialog(elem));
+    window.addEventListener('ak-blanket-click', closeDialog(elem));
+  },
+  detached(elem) {
+    keyPress.destroy();
+    window.removeEventListener('ak-blanket-click', closeDialog(elem));
+  },
   render(elem) {
     const styles = {};
     if (elem.boxShadow) {
@@ -27,24 +54,36 @@ const definition = {
       styles.borderRadius = elem.borderRadius;
     }
 
+    if (elem.open === true) {
+      emit(elem, 'ak-after-open');
+    } else if (elem.open === false) {
+      emit(elem, 'ak-after-close');
+    }
+
     return (
-      <ak-layer
-        open={elem.open}
-        position={elem.position}
-        attachment={elem.constrain}
-        target={elem.target}
-        onRender={(layer) => {
-          if (elem.open && layer.alignment) {
-            elem.positioned = true;
-          }
-        }
-        }
-      >
-        <style>{shadowStyles.toString()}</style>
-        <div class={shadowStyles.locals.inlineDialogContainer} style={styles}>
-          <slot />
-        </div>
-      </ak-layer>
+      <div>
+        {renderBlanketIfNeeded(elem)}
+        <Layer
+          open={elem.open}
+          position={elem.position}
+          attachment={elem.constrain}
+          target={elem.target}
+          onRender={(layer) => {
+            if (elem.open && layer.alignment) {
+              // by default the dialog has opacity 0
+              // and only with attribute 'positioned' it has opacity 1
+              // this behavior is to avoid 'flashing' of a dialog
+              // when it's initially positioning itself on a page
+              elem.setAttribute('positioned', true);
+            }
+          }}
+        >
+          <style>{shadowStyles.toString()}</style>
+          <div className={shadowStyles.locals.inlineDialogContainer} style={styles}>
+            <slot />
+          </div>
+        </Layer>
+      </div>
     );
   },
   props: {
@@ -83,7 +122,6 @@ const definition = {
      */
     open: prop.boolean({
       attribute: true,
-      default: false,
     }),
     /**
      * @description Target of an inline-dialog.
@@ -143,11 +181,55 @@ const definition = {
     padding: prop.string({
       attribute: true,
     }),
-
-    positioned: prop.boolean({
+    /**
+     * @description If dialog has a blanket underneath or not. By default it has.
+     * @memberof InlineDialog
+     * @instance
+     * @type Boolean
+     * @default true
+     * @example @html <ak-inline-dialog has-blanket="true"></ak-inline-dialog>
+     * @example @js dialog.hasBlanket = true
+     */
+    hasBlanket: prop.boolean({
+      attribute: true,
+      default: true,
+    }),
+    /**
+     * @description If click on the blanket dismisses the dialog. By default it is.
+     * @memberof InlineDialog
+     * @instance
+     * @type Boolean
+     * @default true
+     * @example @html <ak-inline-dialog is-blanket-clickable="true"></ak-inline-dialog>
+     * @example @js dialog.isBlanketClickable = true
+     */
+    isBlanketClickable: prop.boolean({
+      attribute: true,
+      default: true,
+    }),
+    /**
+     * @description Is blanket grey with opacity or transparent. By default it's transparent.
+     * @memberof InlineDialog
+     * @instance
+     * @type Boolean
+     * @default false
+     * @example @html <ak-inline-dialog is-blanket-tinted="true"></ak-inline-dialog>
+     * @example @js dialog.isBlanketTinted = true
+     */
+    isBlanketTinted: prop.boolean({
+      attribute: true,
+    }),
+    /**
+     * @description Is blanket is closable by pressing the 'escape' button. By default it is.
+     * @memberof InlineDialog
+     * @instance
+     * @type Boolean
+     * @default true
+     * @example @html <ak-inline-dialog is-closable-on-esc="true"></ak-inline-dialog>
+     * @example @js dialog.isClosableOnEsc = true
+     */
+    isClosableOnEsc: prop.boolean({
       attribute: true,
     }),
   },
-};
-
-export default define('ak-inline-dialog', definition);
+});