+++ packages/pf-mention/src/wc/pf-mention-picker.js
@@ -1,13 +1,11 @@
 import 'style!../host.less';
 import shadowStyles from './pf-mention-picker-shadow.less';
-import pfResourcedMentionList from './pf-resourced-mention-list'; // eslint-disable-line no-unused-vars, max-len
 import { localProp } from './skate-local-props';
+import { define, vdom, prop, props } from 'skatejs';
+import InlineDialog from 'ak-inline-dialog';
+import ResourcedMentionList from './pf-resourced-mention-list';
 
-import 'ak-inline-dialog';
-import { define, vdom, prop, emit, state } from 'skatejs'; // eslint-disable-line no-unused-vars
-
-const definition = {
-
+export default define('pf-mention-picker', {
   prototype: {
     selectNext() {
       if (this._mentionListRef) {
@@ -28,7 +26,7 @@ const definition = {
     },
 
     _filterChange(mentions) {
-      state(this, {
+      props(this, {
         visible: mentions.length > 0,
       });
     },
@@ -49,7 +47,7 @@ const definition = {
   },
 
   render(elem) {
-    const { target, position, renderElementTo } = elem;
+    const { target, position } = elem;
     const { resourceProvider, presenceProvider, query } = elem;
     const style = {
       display: elem.visible ? 'block' : 'none',
@@ -59,20 +57,19 @@ const definition = {
       return (
         <div style={style}>
           <style>{shadowStyles.toString()}</style>
-          <ak-inline-dialog
+          <InlineDialog
             target={target}
             position={position}
             open={elem.visible}
-            renderElementTo={renderElementTo}
             padding="0"
           >
-            <pf-resourced-mention-list
+            <ResourcedMentionList
               resourceProvider={resourceProvider}
               presenceProvider={presenceProvider}
               query={query}
               ref={(ref) => { elem._mentionListRef = ref; }}
             />
-          </ak-inline-dialog>
+          </InlineDialog>
         </div>
       );
     }
@@ -96,6 +93,7 @@ const definition = {
     query: prop.string({
       attribute: true,
     }),
+
     // ak-inline-dialog
     target: prop.string({
       attribute: true,
@@ -103,13 +101,5 @@ const definition = {
     position: prop.string({
       attribute: true,
     }),
-    renderElementTo: prop.string({
-      attribute: true,
-    }),
   },
-};
-
-/* The constructor for our component */
-export default define('pf-mention-picker', definition);
-
-export { definition };
+});