+++ packages/ak-editor-content/CHANGELOG.md
@@ -1,5 +1,15 @@
-<a name="2.0.0"></a>
-# 2.0.0 (2016-07-26)
+<a name="2.0.1"></a>
+## 2.0.1 (2016-08-12)
+
+
+
+<a name="ak-editor-content@2.0.1"></a>
+## ak-editor-content@2.0.1 (2016-08-08)
+
+
+### Bug Fixes
+
+* Disable ::slotted(*) styling until a fix to avoid rule merging is found. ([0e835c3](https://bitbucket.org/atlassian/atlaskit/commits/0e835c3))
 
 
 
@@ -10,6 +20,36 @@
 ### Bug Fixes
 
 * **package:** Add ak- prefix to editor-*, and fix licensing. ([15751dc](https://bitbucket.org/atlassian/atlaskit/commits/15751dc))
+
+
+
+<a name="0.0.1-atlas-animation3"></a>
+## 0.0.1-atlas-animation3 (2016-06-08)
+
+
+
+<a name="1.1.6-atlas-animation"></a>
+## 1.1.6-atlas-animation (2016-06-08)
+
+
+
+<a name="13.1.0-tmp-atlaskit-component-b"></a>
+# 13.1.0-tmp-atlaskit-component-b (2016-06-08)
+
+
+
+<a name="13.1.0-tmp-atlaskit-component-a"></a>
+# 13.1.0-tmp-atlaskit-component-a (2016-06-08)
+
+
+### Bug Fixes
+
+* **components:** commented out a line ([4b91174](https://bitbucket.org/atlassian/atlaskit/commits/4b91174))
+
+
+### Features
+
+* **components:** Add an ! mark ([2b7bdb0](https://bitbucket.org/atlassian/atlaskit/commits/2b7bdb0))