+++ packages/akutil-shared-styles/CHANGELOG.md
@@ -1,5 +1,15 @@
-<a name="33.0.2"></a>
-## 33.0.2 (2016-07-26)
+<a name="33.1.0"></a>
+# 33.1.0 (2016-08-12)
+
+
+
+<a name="akutil-shared-styles@33.1.0"></a>
+# akutil-shared-styles@33.1.0 (2016-08-04)
+
+
+### Features
+
+* Add ak-editor-hyperlink-popup-text package ([e613a16](https://bitbucket.org/atlassian/atlaskit/commits/e613a16))
 
 
 
@@ -10,6 +20,41 @@
 ### Bug Fixes
 
 * **package:** Add ak- prefix to editor-*, and fix licensing. ([15751dc](https://bitbucket.org/atlassian/atlaskit/commits/15751dc))
+
+
+
+<a name="akutil-shared-styles@1.0.0"></a>
+# akutil-shared-styles@1.0.0 (2016-06-24)
+
+
+
+<a name="0.0.1-atlas-animation3"></a>
+## 0.0.1-atlas-animation3 (2016-06-08)
+
+
+
+<a name="1.1.6-atlas-animation"></a>
+## 1.1.6-atlas-animation (2016-06-08)
+
+
+
+<a name="13.1.0-tmp-atlaskit-component-b"></a>
+# 13.1.0-tmp-atlaskit-component-b (2016-06-08)
+
+
+
+<a name="13.1.0-tmp-atlaskit-component-a"></a>
+# 13.1.0-tmp-atlaskit-component-a (2016-06-08)
+
+
+### Bug Fixes
+
+* **components:** commented out a line ([4b91174](https://bitbucket.org/atlassian/atlaskit/commits/4b91174))
+
+
+### Features
+
+* **components:** Add an ! mark ([2b7bdb0](https://bitbucket.org/atlassian/atlaskit/commits/2b7bdb0))