+++ packages/pf-mention/CHANGELOG.md
@@ -1,5 +1,20 @@
-<a name="1.2.0"></a>
-# 1.2.0 (2016-07-26)
+<a name="1.2.2"></a>
+## 1.2.2 (2016-08-12)
+
+
+
+<a name="pf-mention@1.2.2"></a>
+## pf-mention@1.2.2 (2016-08-12)
+
+
+
+<a name="pf-mention@1.2.1"></a>
+## pf-mention@1.2.1 (2016-08-05)
+
+
+### Bug Fixes
+
+* **package:** Fix package dependencies on github packages that were removed ([1c1a99e](https://bitbucket.org/atlassian/atlaskit/commits/1c1a99e))
 
 
 
@@ -57,6 +72,36 @@
 
 <a name="pf-mention@1.0.0"></a>
 # pf-mention@1.0.0 (2016-07-12)
+
+
+
+<a name="0.0.1-atlas-animation3"></a>
+## 0.0.1-atlas-animation3 (2016-06-08)
+
+
+
+<a name="1.1.6-atlas-animation"></a>
+## 1.1.6-atlas-animation (2016-06-08)
+
+
+
+<a name="13.1.0-tmp-atlaskit-component-b"></a>
+# 13.1.0-tmp-atlaskit-component-b (2016-06-08)
+
+
+
+<a name="13.1.0-tmp-atlaskit-component-a"></a>
+# 13.1.0-tmp-atlaskit-component-a (2016-06-08)
+
+
+### Bug Fixes
+
+* **components:** commented out a line ([4b91174](https://bitbucket.org/atlassian/atlaskit/commits/4b91174))
+
+
+### Features
+
+* **components:** Add an ! mark ([2b7bdb0](https://bitbucket.org/atlassian/atlaskit/commits/2b7bdb0))